backend/requirements.txt
aerich==0.7.1
asyncpg
bcrypt
fastapi
passlib
python-jose
tortoise
tortoise-orm==0.19.2
uvicorn